@@ -1,4 +1,4 @@
-class CreatePositiveWords < ActiveRecord::Migration
+class CreatePositiveWords < ActiveRecord::Migration[5.1]
   def change
     create_table :positive_words do |t|
       t.text :title
